Make wasp rows and panels reusabe in other dashboards
