+++ dashboard/panels.go
@@ -1,0 +1,111 @@
+package dashboard
+
+import (
+	"github.com/K-Phoen/grabana/row"
+	"github.com/K-Phoen/grabana/target/prometheus"
+	"github.com/K-Phoen/grabana/timeseries"
+	"github.com/K-Phoen/grabana/timeseries/axis"
+)
+
+func RPSPanel(dataSource string) row.Option {
+	return row.WithTimeSeries(
+		"Responses/sec (Generator, CallGroup)",
+		timeseries.Legend(timeseries.Hide),
+		timeseries.Transparent(),
+		timeseries.Span(6),
+		timeseries.Height("300px"),
+		timeseries.DataSource(dataSource),
+		timeseries.Axis(
+			axis.Unit("Responses"),
+			axis.Label("Responses"),
+		),
+		timeseries.Legend(timeseries.Bottom),
+		timeseries.WithPrometheusTarget(
+			`sum(count_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"responses", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}", call_group=~"${call_group:pipe}"} [1s])) by (node_id, go_test_name, gen_name, call_group)`,
+			prometheus.Legend("{{go_test_name}} {{gen_name}} {{call_group}} responses/sec"),
+		),
+		timeseries.WithPrometheusTarget(
+			`sum(count_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"responses", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"} [1s])) by (node_id, go_test_name, gen_name)`,
+			prometheus.Legend("{{go_test_name}} Total responses/sec"),
+		),
+	)
+}
+
+func RPSVUPerScheduleSegmentsPanel(dataSource string) row.Option {
+	return row.WithTimeSeries(
+		"RPS/VUs per schedule segments",
+		timeseries.Transparent(),
+		timeseries.Span(6),
+		timeseries.Height("300px"),
+		timeseries.DataSource(dataSource),
+		timeseries.WithPrometheusTarget(
+			`
+			max_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"stats", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"}
+			| json
+			| unwrap current_rps [$__interval]) by (node_id, go_test_name, gen_name)
+			`, prometheus.Legend("{{go_test_name}} {{gen_name}} RPS"),
+		),
+		timeseries.WithPrometheusTarget(
+			`
+			sum(last_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"stats", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"}
+			| json
+			| unwrap current_rps [$__interval]) by (node_id, go_test_name, gen_name))
+			`,
+			prometheus.Legend("{{go_test_name}} Total RPS"),
+		),
+		timeseries.WithPrometheusTarget(
+			`
+			max_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"stats", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"}
+			| json
+			| unwrap current_instances [$__interval]) by (node_id, go_test_name, gen_name)
+			`, prometheus.Legend("{{go_test_name}} {{gen_name}} VUs"),
+		),
+		timeseries.WithPrometheusTarget(
+			`
+			sum(last_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"stats", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"}
+			| json
+			| unwrap current_instances [$__interval]) by (node_id, go_test_name, gen_name))
+			`,
+			prometheus.Legend("{{go_test_name}} Total VUs"),
+		),
+	)
+}
+
+func TotalSuccessRequestsPanel(dataSource string) row.Option {
+	return defaultStatWidget(
+		"Successful requests (Total)",
+		dataSource,
+		`
+		sum(max_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"stats", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"}
+		| json
+		| unwrap success [$__range]) by (node_id, go_test_name, gen_name)) by (__stream_shard__)
+		`,
+		`{{go_test_name}} {{gen_name}} Successful requests`,
+	)
+}
+
+func TotalErrorRequestsPanel(dataSource string) row.Option {
+	return defaultStatWidget(
+		"Errored requests (Total)",
+		dataSource,
+		`
+		sum(max_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"stats", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"}
+		| json
+		| unwrap failed [$__range]) by (node_id, go_test_name, gen_name)) by (__stream_shard__)
+		`,
+		`{{go_test_name}} {{gen_name}} Errored requests`,
+	)
+}
+
+func TotalTimedOutRequestsPanel(dataSource string) row.Option {
+	return defaultStatWidget(
+		"Timed out requests (Total)",
+		dataSource,
+		`
+		sum(max_over_time({go_test_name=~"${go_test_name:pipe}", test_data_type=~"stats", branch=~"${branch:pipe}", commit=~"${commit:pipe}", gen_name=~"${gen_name:pipe}"}
+		| json
+		| unwrap callTimeout [$__range]) by (node_id, go_test_name, gen_name)) by (__stream_shard__)
+		`,
+		`{{go_test_name}} {{gen_name}} Timed out requests`,
+	)
+}
